pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "pymupdf>=1.26.5",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sentry-sdk[fastapi]>=2.19.0",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
